fix(underglow): Do not save on/off state during auto state change
When CONFIG_ZMK_RGB_UNDERGLOW_AUTO_OFF_IDLE is enabled, the RGB underglow
turns off and on every time the keyboard goes idle or becomes active.
However, this causes the on/off state of the underglow to be saved to the
storage frequently, which can lead to unnecessary write operations.

To address this issue, this commit separates `zmk_rgb_underglow_on` and
`zmk_rgb_underglow_off` into non-persistent versions that can be used in
`rgb_underglow_auto_state`.

Signed-off-by: XiNGRZ <hi@xingrz.me>